requirements.txt
Django==5.0.1
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
cloudinary==1.40.0
drf-yasg==1.21.7
django-cors-headers==4.3.1
django-filter==24.1
whitenoise==6.6.0
Pillow==10.2.0
gunicorn==21.2.0